Move Socket class into its own file
